refactor(authentication): modify user controller
