+++ app/specs/services/users/users.spec.js
@@ -1,0 +1,7 @@
+// Be descriptive with the what the test case does
+describe('Users factory', function() {
+  it('has a dummy spec to test 2 + 2', function() {
+    // Just one passing test
+    expect(4).toEqual(4);
+  });
+});
