+++ config/express.js
@@ -48,7 +48,7 @@ module.exports = function(app, passport, mongoose) {
             store: new mongoStore({
                 url: config.db,
                 collection: 'sessions',
-                mongoose_connection: mongoose.connection
+                mongooseConnection: mongoose.connection
             })
         }));
