+++ package.json
@@ -9,8 +9,8 @@
     "url": "https://github.com/cardsforhumanity/cardsforhumanity.git"
   },
   "engines": {
-    "node": "0.10.x",
-    "npm": "1.2.x"
+    "node": "8.x.x",
+    "npm": "5.x.x"
   },
   "scripts": {
     "start": "node node_modules/gulp/bin/gulp",
@@ -20,6 +20,7 @@
     "coveralls": "npm run cover -- test/**/ --report lcovonly && cat ./coverage/lcov.info | coveralls"
   },
   "dependencies": {
+    "angular-mocks": "^1.6.6",
     "async": "~0.2.9",
     "bcryptjs": "~1.0.4",
     "bower": "~1.3.8",
@@ -73,11 +74,16 @@
     "grunt-contrib-watch": "~0.5.3",
     "grunt-mocha-test": "~0.7.0",
     "grunt-nodemon": "~0.1.1",
+    "jasmine-core": "^2.8.0",
+    "karma": "^1.7.1",
+    "karma-chrome-launcher": "^2.2.0",
+    "karma-jasmine": "^1.1.0",
+    "karma-mocha": "^1.3.0",
+    "nyc": "^11.2.1",
     "gulp": "^3.9.1",
     "gulp-coveralls": "^0.1.4",
     "istanbul": "^0.4.5",
     "mocha-lcov-reporter": "^1.3.0",
-    "nyc": "^11.2.1",
     "should": "~1.3.0",
     "supertest": "~0.8.0"
   }
